Ability for an OPTIONAL footer.html added targets.
For my own selfish reasons I present a 100% backwards compatible
addition to the StoryFormats: 
If a footer.html file is available in the selected target then it is
used in the output. 
Otherwise, if a footer.html file is not available (like in the old
formats) then the old behaviour still occurs: </div></body></html>

Also added a TEST storyformat to show how this works. This TEST
StoryFormat will not appear in the exe or installer builds.

Change-Id: Ic27867693f9f93e6686a7ca92746e32f4793db60
Signed-off-by: stormrose <eturnerx@gmail.com>