Using File|Close on last story does not exit app
This change in behaviour requested by Philip Sutton. Either the menu
Close Story or the Ctrl-W shortcut key used on the last open story
window will not close the app, instead a new blank story window will
open. Using the X in the upper corner of the window on the last open
story window will exit the app.

Change-Id: I40ccfd20f51a302c5d890b51ad4972e6b7dd2cb4
Signed-off-by: stormrose <eturnerx@gmail.com>