Merge with greyelf
Merging with greyelf's changes.

This also reverts the addition of the 'SugarCube' header - this header
has a number of radical changes which I do not think should be put in
the normal Twine build unconditionally. (I am, however, working on
porting several of its more desirable features to Sugarcane itself.)