Modified Jonah's behaviour
This updates Jonah's idiosyncratic engine in the following manners:
* All passages are now added to the bottom, regardless of where you
click the link.
* Clicking a link to a previously displayed passage will display a new
version of the passage at the bottom, instead of scrolling up to the old
version.
This is based on a patch I've authored based on code by Stormrose
(http://www.glorioustrainwrecks.com/node/5136) and results in more
intuitive and desirable stretchtext behaviour. I'm so confident that
authors would rather have this than Jonah's former TiddlyWiki-esque
behaviour, that I'm not even leaving any trace of the old behaviour
behind.