Updated <<display>> to allow code parameters
* <<display>> now accepts code parameters, such as variables and
functions. <<display $value>>, for instance, can be used.
* Also removed the "_previous" parameter, replacing it with a previous()
function which can be used instead - not just in <<display>> but in
other macros, too.
* Altered visited() to allow multiple parameters. It now returns the
minimum visited value of all parameters (and thus "visited('a','b') can
function as a shorthand for "visited('a') and visited('b')").
* Added either(), a simple randomness function that can be used in
several macros, and is, I feel, more readily adaptable than a randomness
macro. It returns one of its parameters, randomly.