Variables in the link syntax
You can now use variables and expressions in the link syntax, such as
[[$door]] or [[Climb|"Climb" + $var]]. This works like so: if the
passage name is not found, it tries to parse it as a code parameter. If
this successfully returns a value, it is used as the passage name. This
is not a change I am making lightly, but I believe it to be the best
course of action.

* The syntax highlighter now tries to recognise code parameter links. It
currently just looks for $variables, or functions(), but is (much like
with external links) not comprehensive.
* Now, links in the image syntax are recognised as valid outgoing links
in the IDE.
* Altered the variable regexps to exclude all-digit variable names.