An alternative Tag CSS approach
Now, taking a stylesheet passage and adding an extra tag to it (in
addition to "stylesheet") will cause that stylesheet to only be applied
for passages with that given tag.

This, I feel, is a bit tidier than the old Tag CSS method - the
'[data-tags]' selector need not be repeated dozens of times in a
stylesheet, and applying only to the 'body' or '.passage' selectors.
Although, it is slightly less flexible w/r/t tag hierarchies - the
':not[data-tags]' selector can't really be replicated.

The CSS hooks for the old method will remain for
backwards-compatibility, as they are relatively unobtrusive.