Embedded images now usable in stylesheets
Now, using the [img] syntax inside a stylesheet, in places where url()
would be used, will convert it to the CSS url() syntax at runtime.
Importantly, this allows embedded image assets to be usable in
stylesheets - their sole deficiency resolved.

Of course, this means that the stylesheet code is now preprocessed, and
should no longer be assumed to be fully valid CSS outright.