Added parameter() for shorthand <<display>> calls
You can now supply additional arguments to the shorthand <<display>>
macro - in effect, you can "call" passages as if they were macros. (This
is inspired by the "widget" functionality present in Thomas M. Edwards's
SugarCube format.) The parameter() function returns one of the
space-separated terms in the macro invocation, parsed to a value. The
function's argument is a zero-based index - parameter(0) returns the
first parameter, parameter(1) the second, and so forth.

Previously I'd considered a specific "widget" passage type that would
solely permit this type of shorthand call. However, with the advent of
shorthand <<display>> syntax, the necessary distinction between
"widgets" and ordinary passages is now almost absent.