Jonah and <<choice>> changes
* Clicking a link in a past passage in Jonah now rewinds the game state
to that passage before following the link. This gives the author a
greater level of control over the game state - it's no longer possible
to re-use old links from earlier in the story that no longer make sense
in the current context. And, the act of using older links is now more
like "exploring story branches", in that the current branch of the story
is forgotten in favour of the new link's branch. All in all, I believe
it's a good change.
* <<choice>> is now a proper macro in Sugarcane. Instead of merely
creating a link, it now creates a link that can only be clicked once. In
that respect, it now functions like a more atomic <<actions>> macro.