Added variable-setting to the link syntax
Now, you can write [[link text|passage name][$var = "value"]] to make a
passage or external link that changes a variable when it's followed. The
code section is treated as if it were arguments to a <<set>> macro. This
syntax addition should work with <<choice>> macros (but currently not
the [img] syntax).