Made the continued line syntax less strict
Now, the continued line syntax will still work even if some lone
whitespace separates it and the line break. This is a deviation from the
C tradition, but it should save some confusion.

Also made the parameter() variable coerce a falsy argument to 0, so you
can just write parameter() to receive the first argument.

Also tweaked the README.