<<radio>>, <<checkbox>>, <<button>>
* Now, <<textinput>> values are only saved if a <button> link is
pressed. Other links discard the values.
* Added <<radio>>. Its arguments are a variable, and a string for the
label. Radio buttons for the same variable are part of the same set. The
label of the selected radio button is put into the variable.
* Added <<checkbox>>. It's similar to <<radio>>, having identical syntax
- but it sets the variable to an array value containing all of the
checked boxes' labels as strings.
* Added <<button>>, a macro that takes a link as its argument and
creates a <button> element. Use it for passages which use <<radio>> or
<<checkbox>>, or a <<textinput>> set.
* Tried to remove a few potential memory leaks.
* Altered Array.prototype.toString slightly, so that people <<print>>ing
arrays get something that makes it a weensy bit more obvious that it's
an array.
* Updated jQuery to 1.11.