Avoid localized month name in build date
A localized month name can contain non-ASCII characters, which will be
inserted in the string in a locale-specific encoding by strftime().
This can lead to problems: see this forum thread for an example where
an a-umlaut in a Swedish month name led to a build failure:
  http://twinery.org/forum/index.php/topic,756.0.html

I picked the ISO 8601 date format for the numeric-only date.