Removed arguments from Tiddler.update()
The arguments includeInternal and includeMacros were supposed to allow
the caller to select which links should be included. In practice, none
of the callers pass any arguments, so the defaults are always used.

There was a bug in handling wiki-style links inside <<choice>> macros:
the "m" match object was used, which was set in an earlier loop,
instead of the "item" match object that contains the <<choice>> match.
This didn't lead to any visible problems because the regular link
finder would also find all wiki-style links inside macros.

Since it would require extra effort to ensure that when includeInternal
is False the update() method wouldn't pick up wiki-style links inside
macros, I decided it was better to instead drop the includeInternal and
includeMacros arguments.

I did an additional cleanup in parsing the old style <<choice>>
argument syntax: it would strip off text before the pipe character,
but that is relevant only in wiki syntax.