Use sets for storing *_PASSAGES and *_TAGS categories
Since the most common operation on these is a membership test and the
order doesn't matter, a set is a good choice for the data type.

Also use INFO_TAGS in isStoryText() instead of calling several methods
to check for specific tags.

In StoryFrame.buildIncludes(), when including files with the ".tws"
extension, passages with a title that is a *substring* of an excluded
passage name were skipped. I assume this was a bug and the correct
behavior is to skip only passages where the title is *equal* to an
excluded passage name, as was done for included Twee files (albeit
in a slightly overcomplex way).