Optimized the concatenation of story HTML fragments
The main thing is to avoid adding to a string repeatedly, since this
requires an increasing number of bytes to be copied again and again
as the string grows. Instead, put the fragments in a list and at the
end join that list.

Avoid repeated lookup of the tiddler in the dictionary.

Test whether obfuscation is enabled before testing the title: this
avoids the title check completely when not using obfuscation.

The latter two optimizations probably won't have a noticable impact
on performance; the first one probably will on stories with many
passages.