New StorySettings dialog
* The StorySettings passage is now edited in a special GUI dialog, with
checkboxes and textboxes corresponding to each entry.
* The settings for this StorySettings dialog are now stored in the
Header class, in the form of a list of dictionaries. Thus, custom
headers may use different StorySettings entries, if they wish. See
Header.py to examine the current format used for this data.
* Header class: renamed files_to_embed() to filesToEmbed(), primarily
because snake_case just isn't Pythonic(TM).
* Also, the filesToEmbed list now requires that a path be specified,
instead of assuming it's in the format's home path.
* The __init__ method of Headers now takes a third argument:
builtinPath, which should always refer to the path used by the built-in
story formats.
* Added a Help menu to the passage editor, linking to a few wiki
articles concerning passages.