Cleaned up "#!" declarations
Removed them from modules that aren't designed to be executed.
I didn't do a very deep analysis, just common sense and looking for
the existence of a '__name__ == "__main__"' check in the module.

Ask for "python2" as interpreter instead of just "python", see PEP 394
for details.
  http://www.python.org/dev/peps/pep-0394/

Set executable flag on the modules that kept "#!".