Remove latest test build in window close handler instead of destructor
It is hard to predict when destructors run, or even if they run at all:
the documentation states "It is not guaranteed that __del__() methods
are called for objects that still exist when the interpreter exits."
  http://docs.python.org/2/reference/datamodel.html
So it is more reliable to do cleanup actions on window close.

Note that the original code failed to call the superclass destructor.