Monitor clipboard on a timer
This fixes a crash on exit when using wxPython with the GTK backend.
The Clipboard.IsSupported() method seems to cause the event loop to be
re-entered, but if the frame destroy event is handled by that inner
event loop call, a segfault happens. Since the native part of a timer
is removed when the Python proxy reaches 0 references, this problem
does not occur when the clipboard is accessed from a timer.

Note that when the clipboard cannot be opened, the new code will keep
the paste menu entry in unchanged, while the old code would disable
it. I expect the clipboard failing to open to be a temporary condition
in most cases, so then the new behavior makes more sense.