Improved paste implementation in StoryPanel
The old code would forget to close the clipboard if contained data in
an unsupported format. Also it ignored the return value of GetData().

Also renamed the "format" variable since format() is a built-in
function: while it is rarely used, it does receive syntax highlighting
in my editor like any other built-in name, which is distracting.