Fixed a few unterminated RTF control words
In an RTF control word that ends with a digit, a following space is
considered part of the control word. So to make sure spaces in the
following text are not swallowed by the control word, a space should
be output after the digit.

Reported in point 3 of this post:
  http://twinery.org/forum/index.php/topic,1503.0.html