Removed the ObfuscateKey StorySetting
* Removed the 'obfuscatekey' storysetting. Concern has been raised that
the obfuscate option is misleading regarding its purpose: it may lead
authors to assume it is meant as security, rather than as a basic
spoiler protective. So, now it solely uses rot13. HTML files encoded
using an obfuscatekey prior can still be imported.
* Removed String.prototype.unDash.
* Added a rot13() function, used internally to decode rot13 obfuscated
passages, and also of potential use to the author.
* Fixed a bug where the addition of the 'hidden' property to the
storeArea div caused HTML importing to break.
* Fixed some StorySettings dialog glitches.