Forbid assignment = in <<if>> clauses
I want to try making the <<if>> macro print an error if the = sign is
used in one of its clauses, because a bunch of people keep getting bit
by its ambiguity. You might think "well, why not Do What I Mean and
auto-substitute the = for an ==?", but right now I feel doing that would
result in more destructive false-positives than currently, and would of
course encourage the bad habit of not using "is" (or "==" or "===" or
"eq").