Raw HTML passage and imgpassage attributes
One bothersome fact about Twine up to now is that there's no real
"bridge" between the basic Twine link's functionality and raw HTML. The
only working way to make a raw <a> link trigger a passage change is
through a "state.display()" call manually attached to onclick or href.

Now, you can simply make an <a> element (or any element) navigate to a
Twine passage when clicked, by giving it a "passage" attribute - such as
<a passage="Next">. This is converted to an appropriate onclick
attribute at run time. This is confirmed working for both <a> and <area>
links.

Furthermore, for images, an "imgpassage" attribute is available that
lets a raw <img> refer to a Twine image passage - for instance, you can
write <img imgpassage="Tree" passage="Go to tree">.

I've elected not to, as some may propose, extend the "href" and "src"
attributes to support Twine names alongside URLs, because that's bound
to cause more strife than good.