Runtime loading bar
Added a basic loading bar to the story formats, for stories that have a
heavy filesize due to image/font inclusion. It's just a thin coloured
strip on the window-top, in the style of YouTube circa 2014. It updates
based on how many of the #storeArea's nodes have been inserted.

Let no confusion reign: this involves no external file tracking. It is
simply the sole HTML file reporting how much of *itself* has been
downloaded.

Of course, to keep track of this, the runtime needs to know how many
nodes will be in the #storeArea on the outset. So, the story formats'
HTML templates now have a "STORY_SIZE" token that is replaced with the
number of passages in the story at build time.