1.4.2 post-release bugfixes
* Fixed a bug where macros.remember.prefix would always be set to
"Twine.undefined.". Now, it's set to "Twine." plus the default name of
the story (as compared to the StoryTitle name).
* Fixed a bug where the 'cancel' button on offerAssistance dialogs would
still cause the passage to be closed.
* Altered the name of the "Refresh Includes Links" menu item to "Update
StoryIncludes Links". Also, this menu item will be disabled if
StoryIncludes isn't present.
* Now, closing StorySettings will refresh the StoryIncludes links
automatically.
* Fixed a bug where the returned Wikifier.linkFunction would still call
state.display() even when no title was passed (e.g. if a custom macro
created an internal link which, of course, would have this as an event
listener).
* Removed the History.prototype.display startup check, on the basis that
its use-case is a bit too obscure.
* Changed version number to 1.4.3.