Removed 'modified' attribute of tiddlers
This used to contain the modification date, which was used to keep the
newest of two identically titled tiddlers. Instead of keeping track of
the modification date, we now just keep the latest added version. The
only situation when this could lead to different results is when
building a story with some of the tiddlers coming from an HTML file,
which is not a likely scenario.

Note that Tiddler.__cmp__ was implemented incorrectly, but since it was
used incorrectly as well, the two bugs cancelled each other out.