Replaced PassageWidget.sort by posCompare static method
Making it a static method avoid PyLint complaining about the first
argument not being 'self'.

The name was changed because this method itself does not sort anything:
it is a comparator function that can be called by a sorting algorithm.

I also changed the implementation to not compute xDistance unless it
is actually needed. And the final tie breaker was updated to be
antisymmetrical; I think only consistency is needed for Python's sort
implementation, but it was cheap enough to do it properly.