Fix performance issue in Wikifier.parse
The Wikifier.textPrimitives.unquoted method of parsing quoted strings is very slow when dealing with long strings that look like variable names, since it has to test each initial substring to see if it qualifies as "unquoted". This fixes the problem by removing all quoted strings from the input before parsing and restoring them after all of the calls to alter().